requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# AI/GPU Support (Optional but recommended for RTX 5070 Ti)
# Install PyTorch with CUDA 12.1 support:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0
